test(logger): rename all source files to .mjs extension
