+++ src/robots.mjs
