+++ webpack.config.js
@@ -12,13 +12,13 @@
 const path = require('path');
 
 module.exports = {
-  entry: './src/index.js',
+  entry: './src/index.mjs',
   optimization: {
     minimize: false,
   },
   target: 'webworker',
   output: {
-    filename: 'index.js',
+    filename: 'index.mjs',
     path: path.resolve(__dirname, 'bin'),
     libraryTarget: 'this',
   },
